moved ByteFormattingUtil to com.oracle.svm.core.util
