+++ substratevm/src/com.oracle.svm.driver/src/com/oracle/svm/driver/BundleSupport.java
@@ -57,6 +57,7 @@ import java.util.stream.Stream;
 import com.oracle.svm.core.OS;
 import com.oracle.svm.core.option.BundleMember;
 import com.oracle.svm.core.util.ArchiveSupport;
+import com.oracle.svm.core.util.ByteFormattingUtil;
 import com.oracle.svm.driver.BundleOptions.BundleOption;
 import com.oracle.svm.driver.BundleOptions.ExtendedOption;
 import com.oracle.svm.driver.launcher.BundleLauncher;
@@ -64,7 +65,6 @@ import com.oracle.svm.driver.launcher.ContainerSupport;
 import com.oracle.svm.driver.launcher.configuration.BundleArgsParser;
 import com.oracle.svm.driver.launcher.configuration.BundleEnvironmentParser;
 import com.oracle.svm.driver.launcher.configuration.BundlePathMapParser;
-import com.oracle.svm.hosted.ByteFormattingUtil;
 import com.oracle.svm.util.ClassUtil;
 import com.oracle.svm.util.LogUtils;
 import com.oracle.svm.util.StringUtil;
