+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/image/ImageHeapConnectedComponentsPrinter.java
@@ -41,8 +41,8 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import com.oracle.graal.pointsto.BigBang;
+import com.oracle.svm.core.util.ByteFormattingUtil;
 import com.oracle.svm.core.util.VMError;
-import com.oracle.svm.hosted.ByteFormattingUtil;
 import com.oracle.svm.hosted.image.NativeImageHeap.HeapInclusionReason;
 import com.oracle.svm.hosted.image.NativeImageHeap.ObjectInfo;
 import com.oracle.svm.hosted.image.NativeImageHeap.ObjectReachabilityGroup;
